Run softmax in attention dtype (skip float32 cast); total 26.676s→26.669s (0.03%)
Change summary:
- In `core/gemma_forward.py` `AttnScores`, remove the `astype(float32)` cast before softmax so the softmax runs in the native attention dtype.
- This potentially trades numerical stability for speed; intended as a performance experiment only.

Benchmark methodology: 5 full runs of `uv run python -m core.gemma_forward_inference` with the same prompt/settings, reporting mean ± sample std dev for prefill, generation, and total (prefill+generation).

Baseline (mean ± std, n=5): prefill 3.094±0.098s, generation 23.581±0.095s, total 26.676±0.139s.
Variant  (mean ± std, n=5): prefill 3.065±0.053s, generation 23.604±0.104s, total 26.669±0.069s.
Speedup vs baseline: total 26.676s → 26.669s (-0.007s, 0.03%), prefill 3.094s → 3.065s (-0.029s, 0.94%), generation 23.581s → 23.604s (+0.023s, -0.10%).

Net effect is negligible and within noise, and the numeric change is not clearly justified, so this commit is kept only as an experiment record and will be immediately reverted.

+++ core/gemma_forward.py
@@ -133,7 +133,7 @@ def AttnScores(
             idx_a - seq_indices <= config.sliding_window, scores, -jnp.inf
         )
 
-    return jax.nn.softmax(scores.astype(jnp.float32)).astype(scores.dtype)
+    return jax.nn.softmax(scores).astype(scores.dtype)
 
 
 def attnHead(Ks, Vs, Qs, pos_a, is_local_attn) -> jax.Array:
